Move LogManager to log package
